Restructure directories (#19)
