+++ dmc/configs/__init__.py
@@ -2,6 +2,7 @@ from .fig3_robust_sensorimotor_inference import CONFIGS as FIG_3_CONFIGS
 from .fig4_rapid_inference_with_voting import CONFIGS as FIG_4_CONFIGS
 from .fig5_rapid_inference_with_model_based_policies import CONFIGS as FIG_5_CONFIGS
 from .fig6_rapid_learning import CONFIGS as FIG_6_CONFIGS
+from .fig7_flops import CONFIGS as FIG_7_CONFIGS
 from .fig8_multimodal_transfer import CONFIGS as FIG_8_CONFIGS
 from .fig9_structured_object_representations import CONFIGS as FIG_9_CONFIGS
 from .pretraining_experiments import CONFIGS as PRETRAINING_CONFIGS
@@ -13,6 +14,7 @@ CONFIGS.update(FIG_3_CONFIGS)
 CONFIGS.update(FIG_4_CONFIGS)
 CONFIGS.update(FIG_5_CONFIGS)
 CONFIGS.update(FIG_6_CONFIGS)
+CONFIGS.update(FIG_7_CONFIGS)
 CONFIGS.update(FIG_8_CONFIGS)
 CONFIGS.update(FIG_9_CONFIGS)
 CONFIGS.update(VIEW_FINDER_CONFIGS)
