+++ dmc/configs/fig4_rapid_inference_with_voting.py
