+++ dmc/configs/view_finder_images.py
