+++ dmc/scripts/render_view_finder_images.py
